Historical and Geographical Provinces of georgia
User provides as an input  historical provinces of Georgia , which is my country and I really love and appreciate. Program itself checks whether it is one of Georgian states. counts them and if user misses some provinces  creates automatically .csv file that consists of all of the provinces have missed while providing them as an input